+++ tests/test_v2/conftest.py
@@ -2,7 +2,6 @@ import asyncio
 import logging
 import platform
 
-import uvloop
 from pytest import fixture
 
 from cats.v2 import Auth, Handshake, SHA256TimeHandshake
@@ -15,8 +14,6 @@ logging.basicConfig(level='DEBUG', force=True)
 def event_loop():
     if platform.system() == 'Windows':
         asyncio.set_event_loop_policy(asyncio.WindowsSelectorEventLoopPolicy())
-    else:
-        uvloop.install()
     yield asyncio.new_event_loop()